requirements.txt
streamlit==1.36.0
tensorflow==2.16.2
keras==3.4.1
numpy==1.26.4
pillow==10.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1